% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{edgeListToCharVec}
\alias{edgeListToCharVec}
\title{Convert edge list to characteristic vector.}
\usage{
edgeListToCharVec(edgelist, n = NULL)
}
\arguments{
\item{edgelist}{[\code{matrix(2, k)}]
Matrix of edges (each column is one edge).}

\item{n}{[\code{integer}]\cr
Number of nodes of the problem.}
}
\value{
[integer] Characteristic vector cv with cv[i] = 1 if the i-th edge is in the tree.
}
\description{
Convert edge list to characteristic vector.
}
\seealso{
Other transformation functions: \code{\link{permutationToCharVec}},
  \code{\link{permutationToEdgelist}},
  \code{\link{prueferToCharVec}},
  \code{\link{prueferToEdgeList}}
}
