% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{edgeListToGraph}
\alias{edgeListToGraph}
\title{Convert edge list to C++ graph.}
\usage{
edgeListToGraph(edgelist, g)
}
\arguments{
\item{edgelist}{[\code{matrix(2, k)}]
Matrix of edges (each column is one edge).}

\item{[\code{Graph}]\cr}{Object of type graph.}
}
\value{
[\code{Graph}]
}
\description{
Convert edge list to C++ graph.
}
\seealso{
Other transformation functions: \code{\link{charVecToEdgelist}},
  \code{\link{edgeListToCharVec}},
  \code{\link{nodelistToEdgelist}},
  \code{\link{permutationToCharVec}},
  \code{\link{permutationToEdgelist}},
  \code{\link{prueferToCharVec}},
  \code{\link{prueferToEdgeList}}
}
\concept{transformation functions}
