% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoa_Zhou.R
\name{emoaMST_Zhou}
\alias{emoaMST_Zhou}
\title{Pruefer-EMOA for the multi-objective MST problem.}
\usage{
emoaMST_Zhou(instance, n.objectives = 2L, mu, lambda = mu,
  mut = mutUniform2, selMating = ecr::selSimple,
  selSurvival = ecr::selNondom, ref.point, max.iter = 100L)
}
\arguments{
\item{instance}{[\code{any}]\cr
Multi-objective graph problem.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives.}

\item{mu}{[\code{integer(1)}]\cr
Population size.}

\item{lambda}{[\code{integer(1)}]\cr
Number of offspring generated in each generation.
Default is \code{mu}.}

\item{mut}{[\code{ecr_mutator}]\cr
Mutation operator.
Defaults to \code{\link{mutUniform2}}, i.e., each digit of the Pruefer encoding
is replaced with some probability with a random number from \eqn{V = \{1, \ldots, n\}}.}

\item{selMating}{[\code{ecr_selector}]\cr
Mating selector.
Default is \code{\link[ecr]{selSimple}}.}

\item{selSurvival}{[\code{ecr_selector}]\cr
Survival selector.
Default is \code{link[ecr]{selNondom}}.}

\item{ref.point}{[\code{numeric(n.objectives)}]\cr
Reference point for Hypervolume computation.}

\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations.
Default is \code{100}.}
}
\description{
Evolutionary multi-objective algorithm to solve the
multi-objective minimum spanning tree problem. The algorithm adopts the
socalled Pruefer-number as the encoding for spanning trees. A Pruefer-number
for a graph with nodes \eqn{V = \{1, \ldots, n\}} is a sequence of \eqn{n - 2}
number from \eqn{V}. Cayleys theorem states, that a complete graph has
\eqn{n^{n-2}} spanning trees.
The algorithm uses mutation only: each component of an individual is replaced
uniformly at random with another node number from \eqn{V}.
}
\references{
Zhou, G. and Gen, M. Genetic Algorithm Approach on Multi-Criteria
Minimum Spanning Tree Problem. In: European Journal of Operational Research (1999).
}
\seealso{
Other mcMST EMOAs: \code{\link{emoaMST_BG}}
}
