% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerators.R
\name{enumerateTSP}
\alias{enumerateTSP}
\alias{enumerateMST}
\title{Enumerate all solution candidates}
\usage{
enumerateTSP(n)

enumerateMST(n)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Instance size.}
}
\value{
[\code{matrix}] Each row contains a candidate solution.
}
\description{
These functions enumerate all candidate solutions for
a certain combinatorial optimization problem, e.g., all permutations
for a TSP or all Pruefer-codes for a MST problem.
}
\examples{
sols = enumerateTSP(4L)
sols = enumerateMST(4L)
}
