% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genRandomMCGP}
\alias{genRandomMCGP}
\title{Generate a random bi-criteria graph instance.}
\usage{
genRandomMCGP(n)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Instance size, i.e., number of nodes.}
}
\value{
[\code{mcGP}]
}
\description{
No topology is defined. The instance is composed of two
symmetric weight matrices.
}
\examples{
inst = genRandomMCGP(6L)
}
