% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtremeSolutions.R
\name{getExtremeSolutions}
\alias{getExtremeSolutions}
\title{Compute extreme spanning trees of bi-criteria graph problem.}
\usage{
getExtremeSolutions(graph)
}
\arguments{
\item{graph}{[\code{mcGP}]\cr
Multi-objective graph problem.}
}
\value{
[\code{matrix(2, 2)}] The i-th column contains the objective vector
of the extreme i-th extreme solution
}
\description{
Internally \code{\link{mcMSTPrim}} is called with weights set accordingly.
}
