% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcMSTEmoaBG.R
\name{mcMSTEmoaBG}
\alias{mcMSTEmoaBG}
\title{Subgraph EMOA for the multi-criteria MST problem.}
\usage{
mcMSTEmoaBG(instance, n.objectives = 2L, mu, lambda = mu, mut = NULL,
  selMating = NULL, selSurvival = ecr::selNondom, ref.point = NULL,
  max.iter = 100L)
}
\arguments{
\item{instance}{[\code{any}]\cr
Multi-objective graph problem.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives.}

\item{mu}{[\code{integer(1)}]\cr
Population size.}

\item{lambda}{[\code{integer(1)}]\cr
Number of offspring generated in each generation.
Default is \code{mu}.}

\item{mut}{[\code{ecr_mutator}]\cr
Mutation operator.
Default is \code{\link{mutSubgraphMST}}.}

\item{selMating}{[\code{ecr_selector}]\cr
Mating selector.
Default is \code{\link[ecr]{selSimple}}.}

\item{selSurvival}{[\code{ecr_selector}]\cr
Survival selector.
Default is \code{link[ecr]{selNondom}}.}

\item{ref.point}{[\code{numeric(n.objectives) | NULL}]\cr
Reference point for Hypervolume computation.
If \code{NULL} the sum of the \eqn{n} largest edges in each objective
is taken where \eqn{n} is the number of nodes of \code{instance}.
This is an upper bound for the size of each spanning tree
with \eqn{(n-1)} edges.}

\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations.
Default is \code{100}.}
}
\value{
[\code{\link[ecr]{ecr_result}}]
}
\description{
Evolutionary multi-objective algorithm to solve the
multi-objective minimum spanning tree problem. The algorithm relies
to mutation only to generate offspring employing the subgraph mutator
(see \code{\link{mutSubgraphMST}}).
}
\note{
This algorithm performs minimization in all objectives.
}
\examples{
inst = genRandomMCGP(10)
res = mcMSTEmoaBG(inst, mu = 20L, ref.point = c(1e5, 1e5), max.iter = 100L)
print(res$pareto.front)
print(tail(getStatistics(res$log)))
}
\seealso{
Mutators \code{\link{mutSubgraphMST}} and \code{\link{mutEdgeExchange}}

Other mcMST EMOAs: \code{\link{mcMSTEmoaZhou}}

Other mcMST algorithms: \code{\link{mcMSTEmoaZhou}},
  \code{\link{mcMSTPrim}}
}
