% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{network}
\alias{network}
\title{Generate a bare network (only the number of nodes is known).}
\usage{
network(n.dims = 2L, lower, upper)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of nodes.}
}
\value{
[\code{Network}]
}
\description{
Generate a bare network (only the number of nodes is known).
}
