% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{permutationToCharVec}
\alias{permutationToCharVec}
\title{Convert permutation to characteristic vector.}
\usage{
permutationToCharVec(perm, n)
}
\arguments{
\item{perm}{[\code{integer}]\cr
Permutation of nodes, e.g., solution of a TSP.}

\item{n}{[\code{integer}]\cr
Number of nodes of the problem.}
}
\value{
[integer] Characteristic vector cv with cv[i] = 1 if the i-th edge is in the tree.
}
\description{
Convert permutation to characteristic vector.
}
\seealso{
Other transformation functions: \code{\link{edgeListToCharVec}},
  \code{\link{permutationToEdgelist}},
  \code{\link{prueferToCharVec}},
  \code{\link{prueferToEdgeList}}
}
