% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{permutationToEdgelist}
\alias{permutationToEdgelist}
\title{Convert permutation to edge list.}
\usage{
permutationToEdgelist(perm)
}
\arguments{
\item{perm}{[\code{integer}]\cr
Permutation of nodes, e.g., solution of a TSP.}
}
\value{
[\code{matrix(2, length(perm))}] Edge list.
}
\description{
Convert permutation to edge list.
}
\seealso{
Other transformation functions: \code{\link{edgeListToCharVec}},
  \code{\link{permutationToCharVec}},
  \code{\link{prueferToCharVec}},
  \code{\link{prueferToEdgeList}}
}
