% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{prueferToCharVec}
\alias{prueferToCharVec}
\title{Convert Pruefer code to characteristic vector.}
\usage{
prueferToCharVec(pcode)
}
\arguments{
\item{pcode}{[\code{integer}]
Pruefer code encoding a minimum spanning tree.}
}
\value{
[integer] Characteristic vector cv with cv[i] = 1 if the i-th edge is in the tree.
}
\description{
Convert Pruefer code to characteristic vector.
}
\seealso{
Other transformation functions: \code{\link{edgeListToCharVec}},
  \code{\link{permutationToCharVec}},
  \code{\link{permutationToEdgelist}},
  \code{\link{prueferToEdgelist}}
}
