% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{prueferToEdgelist}
\alias{prueferToEdgelist}
\title{Convert Pruefer code to edge list.}
\usage{
prueferToEdgelist(pcode)
}
\arguments{
\item{pcode}{[\code{integer}]
Pruefer code encoding a minimum spanning tree.}
}
\value{
[\code{matrix(2, length(pcode) + 1)}] Edge list.
}
\description{
Convert Pruefer code to edge list.
}
\seealso{
Other transformation functions: \code{\link{edgeListToCharVec}},
  \code{\link{permutationToCharVec}},
  \code{\link{permutationToEdgelist}},
  \code{\link{prueferToCharVec}}
}
