% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sampleWeights}
\alias{sampleWeights}
\title{Sample weights}
\usage{
sampleWeights(n)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of weights to sample.}
}
\value{
[\code{numeric}] Weight vector.
}
\description{
Sample random weights \eqn{\lambda_1, ... \lambda_n, \sum_{i=1}^{n} \lambda_i = 1}
for weighted-sum scalarization.
}
\examples{
sampleWeights(2)

weights = replicate(10, sampleWeights(3L))
colSums(weights)
}
