% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveMCGP}
\alias{saveMCGP}
\title{Store an mcGP instance to file.}
\usage{
saveMCGP(instance, file)
}
\arguments{
\item{instance}{[\code{mcGP}]\cr
Problem instance.}

\item{file}{[\code{character(1)}]\cr
Path to file.}
}
\value{
[invisible(logical(1))]
}
\description{
Store an mcGP instance to file.
}
