% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scalarizeWeights}
\alias{scalarizeWeights}
\title{Scalarize weight matrizes.}
\usage{
scalarizeWeights(weight.mats, lambdas)
}
\arguments{
\item{weight.mats}{[\code{list}]\cr
List of weight matrizes.}

\item{lambdas}{[\code{numeric}]\cr
Vector of weights.}
}
\value{
[\code{matrix}]
}
\description{
Given a list of weight matrizes \code{weight.mats} and a vector
of numeric weights, the function returns a single weight matrix. Each component
of the resulting matrix is the weighted sum of the corresponding components of
the weight matrizes passed.
}
