% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{addWeights}
\alias{addWeights}
\title{Add weights to a multi-objective graph.}
\usage{
addWeights(graph, method = "euclidean", weight.fun = NULL,
  symmetric = TRUE, ...)
}
\arguments{
\item{graph}{[\code{mcGP}]\cr
Multi-objective graph problem.}

\item{method}{[\code{character(1)}]\cr
Method to generate weights based on node coordinates. See \code{method} argument
of \code{\link[stats]{dist}}.}

\item{weight.fun}{[\code{function(m, ...) | NULL}]\cr
Function used to generate weights. The first arument needs to be number of weights
to generate.}

\item{symmetric}{[\code{logical(1)}]\cr
Should the weights be symmetric, i.e., w(i, j) = w(j, i) for each pair i, j of nodes?
Default is \code{TRUE}.}

\item{...}{[any]\cr
Additional arguments passed down to \code{weight.fun}.}
}
\value{
[\code{mcGP}] Multi-objective graph problem.
}
\description{
\code{addWeights} allows to generate edge weights for a multi-objective
graph instance. The weights can be generated on basis of the node coordinates (in this
case \code{\link[stats]{dist}} is applied with the cooresponding \code{method}).
Alternatively, all kinds of random weights can be generated.
}
