% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{genRandomMCGP}
\alias{genRandomMCGP}
\title{Generate a bi-criteria graph with two uniformly randomly distribted edge weights.}
\usage{
genRandomMCGP(n)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Instance size, i.e., number of nodes.}
}
\value{
[\code{\link[grapherator]{grapherator}}]
}
\description{
The instance is composed of two
symmetric weight matrices. The first weight is drawn independently at
random from a \eqn{\mathcal{R}[10, 100]} distribution, the second one
from a \eqn{\mathcal{R}[10, 50]} distribution (see references).
}
\note{
This is a simple wrapper around the much more flexible graph generation
system in package \pkg{grapherator}.
}
\examples{
g = genRandomMCGP(10L)
\dontrun{
pl = grapherator::plot(g)
}
}
\references{
Zhou, G. and Gen, M. Genetic Algorithm Approach on Multi-Criteria
Minimum Spanning Tree Problem. In: European Journal of Operational Research (1999).

Knowles, JD & Corne, DW 2001, A comparison of encodings and algorithms for multiobjective
minimum spanning tree problems. in Proceedings of the IEEE Conference on Evolutionary
Computation, ICEC|Proc IEEE Conf Evol Comput Proc ICEC. vol. 1, Institute of Electrical
and Electronics Engineers , pp. 544-551, Congress on Evolutionary Computation 2001,
Soul, 1 July.
}
