% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeight.R
\name{getWeight}
\alias{getWeight}
\title{Get the overall costs/weight of a suggraph given its edgelist.}
\usage{
getWeight(instance, edgelist)
}
\arguments{
\item{instance}{[\code{any}]\cr
Multi-obective graph problem.}

\item{edgelist}{[\code{matrix(2, k)}]
Matrix of edges (each column is one edge).}
}
\value{
[numeric(2)] Weight vector.
}
\description{
Get the overall costs/weight of a suggraph given its edgelist.
}
